--- .prettierrc.json
@@ -1,5 +1,0 @@
-{
-  "semi": false,
-  "trailingComma": "es5",
-  "singleQuote": true
-}
