+++ components/demo.module.css
@@ -1,0 +1,4 @@
+.select {
+  width: 100%;
+  padding: 8px 16px;
+}
