+++ package.json
@@ -1,38 +1,34 @@
 {
   "name": "next-qrcode-site",
-  "version": "0.1.0",
-  "private": true,
+  "version": "0.0.1",
+  "description": "A QR code plugin for Next.js, Create React App, Gatsby.js as well as React apps.",
   "scripts": {
     "dev": "next dev",
     "build": "next build",
-    "export": "next export",
-    "start": "next start",
-    "lint": "next lint",
-    "format": "prettier --check --ignore-path .gitignore .",
-    "format:fix": "prettier --write --ignore-path .gitignore ."
+    "start": "next start"
   },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/shuding/nextra-docs-template.git"
+  },
+  "author": "Shu Ding <g@shud.in>",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/shuding/nextra-docs-template/issues"
+  },
+  "homepage": "https://github.com/shuding/nextra-docs-template#readme",
   "dependencies": {
-    "@chakra-ui/react": "^2.3.6",
-    "@emotion/react": "^11.10.4",
-    "@emotion/styled": "^11.10.4",
-    "framer-motion": "^6.5.1",
-    "next": "12.3.1",
-    "next-prism": "^0.3.0",
+    "next": "^13.0.6",
+    "next-prism": "^0.5.0",
     "next-qrcode": "^2.4.1",
-    "react": "18.2.0",
+    "nextra": "latest",
+    "nextra-theme-docs": "latest",
+    "react": "^18.2.0",
     "react-color": "^2.19.3",
-    "react-dom": "18.2.0",
-    "react-select": "^5.5.5"
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
-    "@types/node": "18.11.5",
-    "@types/react": "18.0.22",
-    "@types/react-color": "^3.0.6",
-    "@types/react-dom": "18.0.7",
-    "eslint": "8.26.0",
-    "eslint-config-next": "12.3.1",
-    "eslint-config-prettier": "^8.5.0",
-    "prettier": "^2.7.1",
-    "typescript": "4.8.4"
+    "@types/node": "18.11.10",
+    "typescript": "^4.9.3"
   }
 }
