--- pages/_app.tsx
@@ -1,23 +1,0 @@
-import type { AppProps } from 'next/app'
-import Head from 'next/head'
-
-import { ChakraProvider } from '@chakra-ui/react'
-import 'next-prism/themes/tomorrow.css'
-
-function MyApp({ Component, pageProps }: AppProps) {
-  return (
-    <>
-      <Head>
-        <title>
-          next-qrcode - React hooks for generating QRCode for your next React
-          apps
-        </title>
-      </Head>
-      <ChakraProvider>
-        <Component {...pageProps} />
-      </ChakraProvider>
-    </>
-  )
-}
-
-export default MyApp
