--- pages/_document.tsx
@@ -1,33 +1,0 @@
-import Document, {
-  DocumentContext,
-  Html,
-  Head,
-  Main,
-  NextScript,
-} from 'next/document'
-
-class MyDocument extends Document {
-  static async getInitialProps(ctx: DocumentContext) {
-    const initialProps = await Document.getInitialProps(ctx)
-    return initialProps
-  }
-
-  render() {
-    return (
-      <Html lang="en">
-        <Head>
-          <meta
-            name="description"
-            content="React hooks for generating QRCode for your next React apps"
-          />
-        </Head>
-        <body>
-          <Main />
-          <NextScript />
-        </body>
-      </Html>
-    )
-  }
-}
-
-export default MyDocument
