+++ pages/demo.mdx
@@ -1,0 +1,3 @@
+import Demo from '../components/demo'
+
+<Demo />
