+++ pages/index.mdx
@@ -1,0 +1,17 @@
+# Introduction
+ 
+A QR code plugin for Next.js, Create React App, Gatsby.js as well as React apps.
+
+### Installation
+
+next-qrcode is available on npm. It can be installed with the following command:
+
+```sh copy
+npm install next-qrcode --save
+```
+
+next-qrcode is available on yarn as well. It can be installed with the following command:
+
+```sh copy
+yarn add next-qrcode --save
+```
