+++ pages/use-qrcode/_meta.json
@@ -1,0 +1,5 @@
+{
+  "canvas": "Canvas",
+  "svg": "SVG",
+  "image": "Image"
+}
